added copyright header.  generalised abillity system.
