+++ Assets/MyGameplayAbilitySystem/Abilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d38ac39699fdd1428feff95e33a00c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
