+++ Assets/MyGameplayAbilitySystem/Abilities/DefaultAttackAbility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5b42232cf4782d4ba5e248d7af6a0ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
