+++ Assets/Plugins/GameplayAbilitySystem/Abilities/Systems/AbilitySystem.cs
@@ -1,11 +1,30 @@
-using GameplayAbilitySystem.Abilities.Systems;
+/*
+ * Created on Mon Nov 04 2019
+ *
+ * The MIT License (MIT)
+ * Copyright (c) 2019 Sahil Jain
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ * and associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
+ * TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+ * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 using GameplayAbilitySystem.Common.Components;
 using GameplayAbilitySystem.GameplayEffects.Components;
 using Unity.Burst;
 using Unity.Collections;
 using Unity.Entities;
 using Unity.Jobs;
-using UnityEngine;
 
 namespace GameplayAbilitySystem.Abilities.Systems {
 
@@ -25,6 +44,9 @@ namespace GameplayAbilitySystem.Abilities.Systems {
         protected abstract JobHandle CheckAbilityAvailable(JobHandle inputDeps);
         protected abstract JobHandle CooldownJobs(JobHandle inputDeps);
 
+        /// <summary>
+        /// Gather all cooldown effects for this ability
+        /// </summary>
         [BurstCompile]
         protected struct GatherCooldownGameplayEffectsJob : IJobForEach<GameplayEffectTargetComponent, GameplayEffectDurationComponent> {
             public NativeMultiHashMap<Entity, GameplayEffectDurationComponent>.ParallelWriter GameplayEffectDurations;
@@ -33,6 +55,9 @@ namespace GameplayAbilitySystem.Abilities.Systems {
             }
         }
 
+        /// <summary>
+        /// Get the longest cooldown for the ability for each entity
+        /// </summary>
         [BurstCompile]
         [RequireComponentTag(typeof(AbilitySystemActor))]
         protected struct GatherLongestCooldownPerEntity : IJobForEachWithEntity<T> {
@@ -63,10 +88,14 @@ namespace GameplayAbilitySystem.Abilities.Systems {
             }
         }
 
+        /// <summary>
+        /// If an entity has no active cooldowns active for an ability, we would never get the chance
+        /// to update the cooldown to 0.  This inserts a default cooldown of '0', so we can
+        /// use that as the default for each ability.
+        /// </summary>
         [BurstCompile]
         protected struct CooldownAbilityIsZeroIfAbsentJob : IJobForEachWithEntity<T> {
             public NativeMultiHashMap<Entity, GameplayEffectDurationComponent>.ParallelWriter GameplayEffectDurations;
-
             public void Execute(Entity entity, int index, ref T abilityDurationComponent) {
                 GameplayEffectDurations.Add(entity, GameplayEffectDurationComponent.Initialise(0, 0));
             }
