+++ Assets/Plugins/GameplayAbilitySystem/Attributes/CharacterAttributesScriptableObject/Editor/CharacterAttributesScriptableObjectEditor.cs
@@ -1,10 +1,27 @@
-﻿using System.Collections;
+﻿/*
+ * Created on Mon Nov 04 2019
+ *
+ * The MIT License (MIT)
+ * Copyright (c) 2019 Sahil Jain
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ * and associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
+ * TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+ * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 using System.Collections.Generic;
 using System.Linq;
-using GameplayAbilitySystem;
 using UnityEditor;
-using UnityEditor.UIElements;
-using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace GameplayAbilitySystem.Attributes.Components {
@@ -27,7 +44,7 @@ namespace GameplayAbilitySystem.Attributes.Components {
                     "Assets/Plugins/GameplayAbilitySystem/Attributes/CharacterAttributesScriptableObject/Editor/CharacterAttributesScriptableObjectEditor.uss"
                 );
             m_RootElement.styleSheets.Add(stylesheet);
-            
+
             // Cleanup any strings which correspond to types that no longer exist.
             var allTypes = new AttributeCollector().GetAllAttributeClasses(System.AppDomain.CurrentDomain);
             var serializedTypeStrings = new List<string>();
